Flip around _decimal and decimal
