+++ stdlib/3/_pydecimal.pyi
@@ -1,4 +1,4 @@
 
 # This is a slight lie, the implementations aren't exactly identical
 # However, in all likelihood, the differences are inconsequential
-from _decimal import *
+from decimal import *
